change ModMove to resize from the closest corner
Previously, ModMove always resized the window as if dragging from the bottom
right corner. This change makes it as if the drag event occurs from whatever
corner of the window the mouse is closest to.

Unfortunately, dragging from any other corner than the bottom right is somewhat
less precise, in that the overall position of the window can shift slightly,
especially when dragging faster. I think this is just a consequence of the way
the mouseMoved functions is called over and over, instead of saving the first
and last position.