improve precision/smoothness of move and resize
This change saves the initial conditions from when the gesture is initialized
and computes moves/resizes from that. This avoids accumulating errors from
continuously computing deltas only based on the last observed mouse position.

Now, the relative position of the mouse cursor within the window does not
change, no matter how quickly the mouse is moved. Similarly with resizes, the
non-moving corner of the window stays in exactly the same place (this fixes the
janky behavior that the previous commit resulted in).